feat: remove model dep on config file
